requirements.txt
blinker==1.4
flask==2.0.3
flask-dance==6.0.0
flask-login==0.6.1
flask-sqlalchemy==2.5.1
gunicorn==20.1.0
python-dotenv==0.20.0
setuptools==62.1.0
